pyproject.toml
[project]
name = "discord-voice-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "python-dotenv>=1.1.0",
]
